requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
brotlipy==0.7.0
cachetools==5.3.0

frozenlist==1.3.3
google==3.0.0
google-api-core==2.11.0
google-auth==2.15.0
google-cloud==0.34.0
googleapis-common-protos==1.57.0
google-cloud-texttospeech

imageio==2.25.0
iso8601==1.1.0

multidict==6.0.4
networkx==3.0
numpy==1.24.1
nltk
discord.py
python-dotenv
openai==0.27.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
python-vlc
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8

PyWavelets==1.4.1

rsa==4.9

scipy==1.10.0

soupsieve==2.4
tifffile==2022.10.10


twitchio==2.6.0
typing_extensions==4.4.0


wincertstore==0.2
yarl==1.8.2

python-dotenv
